# packages/api-contract/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/projects/{projectId}/photos": {
      "get": {
        "operationId": "listPhotos",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "每页照片数量",
            "schema": {
              "default": 12,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "name": "kind",
            "required": false,
            "in": "query",
            "description": "照片类型；不传时返回全部照片",
            "schema": {
              "$ref": "#/components/schemas/PhotoKind"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "上一页返回的游标",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "图库项目 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoPageDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "photos"
        ]
      },
      "post": {
        "operationId": "uploadPhoto",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "图库项目 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "description": "要上传的单张原始图片",
                    "format": "binary",
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "413": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "415": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "photos"
        ]
      }
    },
...
    "/material/stickers": {
      "get": {
        "operationId": "listStickers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MaterialStickerDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "material"
        ]
      },
      "post": {
        "operationId": "uploadSticker",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialStickerDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "material"
        ]
      }
    },
    "/material/stickers/{stickerId}": {
      "delete": {
        "operationId": "deleteSticker",
        "parameters": [
          {
            "name": "stickerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "material"
        ]
      }
    },
    "/material/users": {
      "get": {
        "operationId": "listUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "返回所有用户及每个用户用有的图库数量",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MaterialUserDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "material"
        ]
      }
    },
    "/material/users/{userId}/projects": {
      "get": {
        "operationId": "listUserProjects",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "用户 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "material"
        ]
      }
    },
    "/material/projects/{projectId}/photos": {
      "get": {
        "operationId": "listProjectPhotos",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "每页返回的图片数量",
            "schema": {
              "default": 24,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "name": "kind",
            "required": false,
            "in": "query",
            "description": "照片类型；不传时返回全部照片",
            "schema": {
              "$ref": "#/components/schemas/PhotoKind"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "上一页返回的图片游标",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "图库项目 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialPhotoPageDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "material"
        ]
      }
    },
    "/material/favorites": {
      "get": {
        "operationId": "listFavoritePhotos",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "每页返回的图片数量",
            "schema": {
              "default": 24,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "name": "kind",
            "required": false,
            "in": "query",
            "description": "照片类型；不传时返回全部照片",
            "schema": {
              "$ref": "#/components/schemas/PhotoKind"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "上一页返回的图片游标",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialPhotoPageDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "material"
        ]
      }
    },
    "/material/photos/{photoId}/like": {
      "put": {
        "operationId": "likeMaterialPhoto",
        "parameters": [
          {
            "name": "photoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialPhotoLikeStateDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "material"
        ]
      },
      "delete": {
        "operationId": "unlikeMaterialPhoto",
        "parameters": [
          {
            "name": "photoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialPhotoLikeStateDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "material"
        ]
      }
    },
    "/material/photos/{photoId}/favorite": {
      "put": {
        "operationId": "favoriteMaterialPhoto",
        "parameters": [
          {
            "name": "photoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialPhotoFavoriteStateDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "material"
        ]
      },
      "delete": {
        "operationId": "unfavoriteMaterialPhoto",
        "parameters": [
          {
            "name": "photoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialPhotoFavoriteStateDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "material"
        ]
      }
    },
    "/material/photos/{photoId}/comments": {
      "get": {
        "operationId": "listMaterialPhotoComments",
        "parameters": [
          {
            "name": "photoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MaterialCommentDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "material"
        ]
      },
      "post": {
        "operationId": "createMaterialPhotoComment",
        "parameters": [
          {
            "name": "photoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMaterialCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialCommentDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "material"
        ]
      }
    },
    "/material/comments/{commentId}": {
      "delete": {
        "operationId": "deleteMaterialPhotoComment",
        "parameters": [
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "material"
        ]
      }
    },
...
  },
  "info": {
    "title": "Photo Lab API",
    "description": "Generated OpenAPI contract for Photo Lab clients.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBodyDto"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "PhotoKind": {
        "type": "string",
        "enum": [
          "ORIGINAL",
          "EDITED"
        ]
      },
...
      "PhotoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "fileName": {
            "type": "string",
            "description": "用户上传时的原始文件名"
          },
          "mimeType": {
            "type": "string",
            "example": "image/jpeg"
          },
          "originalUrl": {
            "type": "string",
            "description": "原始图片的 API 相对 URL",
            "example": "/public/projects/project-id/photos/photo-id--holiday.jpg"
          },
          "thumbnailUrl": {
            "type": "string",
            "description": "需要认证的按需缩略图 API 相对 URL",
            "example": "/projects/project-id/photos/photo-id/thumbnail"
          },
          "sizeBytes": {
            "type": "number",
            "description": "原始图片大小，单位为字节",
            "example": 4839201
          },
          "width": {
            "type": "number",
            "nullable": true
          },
          "height": {
            "type": "number",
            "nullable": true
          },
          "kind": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PhotoKind"
              }
            ]
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PhotoStatus"
              }
            ]
          },
          "likeCount": {
            "type": "number",
            "description": "图片收到的点赞总数"
          },
          "favoriteCount": {
            "type": "number",
            "description": "图片被收藏的总数，仅图片所有者可见"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "projectId",
          "fileName",
          "mimeType",
          "originalUrl",
          "thumbnailUrl",
          "sizeBytes",
          "width",
          "height",
          "kind",
          "status",
          "likeCount",
          "favoriteCount",
          "createdAt",
          "updatedAt"
        ]
      },
      "PhotoPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PhotoDto"
            }
          },
          "nextCursor": {
            "type": "string",
            "description": "下一页游标；没有更多数据时为 null",
            "nullable": true
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      },
...
      "MaterialStickerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "url",
          "createdAt"
        ]
      },
      "MaterialUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "用户id"
          },
          "userName": {
            "type": "string",
            "description": "用户名"
          },
          "projectCount": {
            "type": "number",
            "description": "用户拥有图库数量",
            "example": 3,
            "minimum": 0
          }
        },
        "required": [
          "id",
          "userName",
          "projectCount"
        ]
      },
...
      "MaterialPhotoFavoriteStateDto": {
        "type": "object",
        "properties": {
          "photoId": {
            "type": "string"
          },
          "isFavorited": {
            "type": "boolean"
          }
        },
        "required": [
          "photoId",
          "isFavorited"
        ]
      },
...
    }
  }
}